+++ src/main/java/roomescape/controller/ThemeController.java
@@ -5,7 +5,6 @@ import java.util.List;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -13,7 +12,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import roomescape.dto.ErrorResponse;
 import roomescape.dto.ThemeCreateRequest;
 import roomescape.dto.ThemeResponse;
 import roomescape.service.ThemeService;
@@ -55,13 +53,6 @@ public class ThemeController {
         return ResponseEntity
                 .noContent()
                 .build();
-    }
-
-    @ExceptionHandler({IllegalArgumentException.class, NullPointerException.class})
-    public ResponseEntity<ErrorResponse> handleException(RuntimeException e) {
-        return ResponseEntity
-                .badRequest()
-                .body(new ErrorResponse(e.getMessage()));
     }
 }
