+++ src/main/java/roomescape/controller/TimeController.java
@@ -5,7 +5,6 @@ import java.util.List;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -15,7 +14,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import roomescape.dto.AvailableTimeResponse;
-import roomescape.dto.ErrorResponse;
 import roomescape.dto.TimeCreateRequest;
 import roomescape.dto.TimeResponse;
 import roomescape.service.TimeService;
@@ -57,13 +55,6 @@ public class TimeController {
         return ResponseEntity
                 .noContent()
                 .build();
-    }
-
-    @ExceptionHandler({IllegalArgumentException.class, NullPointerException.class})
-    public ResponseEntity<ErrorResponse> handleException(RuntimeException e) {
-        return ResponseEntity
-                .badRequest()
-                .body(new ErrorResponse(e.getMessage()));
     }
 }
